Call populatevolumes after compeleting mount
1. Loop over all mounts
2. For each copy existing data
3. Drop a markfile to not repeat the flow
